Bug fix: widgets were being marked dirty during initialization.  More...
() Weblocks does some MOP hacking to cause widgets to be marked dirty
when one of their slots is modified.  But, this was happening even when
the widget was being initialized; causing their RENDER-WIDGET methods to
be called unnecessarily.  This was at least a performance bug.  Now, a
widget is never marked dirty until it has been rendered the first time.

() I undid some changes I had previously made to quiet warnings issued
when strings and functions were used as widgets.  It eventually became
clear this was not a good idea; only widgets should be in the widget tree.
I added a couple of calls to MAKE-WIDGET to better enforce this, and
restored the warnings.  (I also made it an error to pass a string in the
list of children passed to MAKE-WIDGET; an error that's easy to make is
to forget to wrap a call to something like WITH-HTML or RENDER-LINK in a
lambda, and this detects that.)

() I added a :DOM-CLASS argument to MAKE-WIDGET.

() Finally, I added a convenience macro MAKE-FLOW-WIDGET that makes it
easier to use flows.