Support for non-AJAX drilldowns in datagrids.  With these, the browser's "back" button returns from the drilldown widget to the datagrid, as the user might expect.
Example: you can have a datagrid of frobs selected by a 'navigation' at
'http://foo.com/frobs/'; the :drilldown-link-url-fn can return URLs like
'http://foo.com/frobs/42'.  To do this, wrap the datagrid in a custom
subclass of 'on-demand-selector', with a :lookup-function that, if given
an empty list of tokens, returns the datagrid itself, otherwise returns
the drilldown widget named by the first token.