* src/widgets/widget/widget.lisp: fix multiple RENDER-WIDGET from slburson / weblocks
https://github.com/slburson/weblocks/commit/cc290476e165155f0f39386f2fd546fced546390

widgets were being marked dirty during initialization. More...
() Weblocks does some MOP hacking to cause widgets to be marked dirty
when one of their slots is modified.  But, this was happening even when
the widget was being initialized; causing their RENDER-WIDGET methods to
be called unnecessarily.  This was at least a performance bug.  Now, a
widget is never marked dirty until it has been rendered the first time.